karinca_algoritma_muratpasa/requirements.txt
streamlit
googlemaps
numpy
pandas
polyline
folium
streamlit-folium
matplotlib
python-dotenv
